Clean up style files.
