+++ client/components/common/styles/index.js
@@ -1,0 +1,3 @@
+export * from './base';
+export * from './inline';
+export * from './rules';
