+++ client/components/layout/styles/index.js
@@ -1,0 +1,2 @@
+export * from './inline';
+export * from './rules';
