+++ client/components/navigation/styles/index.js
@@ -1,0 +1,1 @@
+export * from './inline';
