+++ client/constants/svg.js
@@ -5,3 +5,7 @@ export const PINGED = 'PINGED';
 export const PING_TIMEOUT = 5000;
 export const PING_ANIMATION_LOOPS = 8;
 export const PING_ANIMATION_TIMEOUT = PING_TIMEOUT / 25;
+
+export const TEXT_DX = 2;
+export const TEXT_DY = 24;
+export const MARKER_TEXT_DY = TEXT_DY + 16;
